Converts a Raster Layer to Single Objective Function

Takes the altitude informations from a RasterLayer Object as an objective for a smoof function.

convertRasterLayerToSmoof(raster.layer, maximize = TRUE, interpolate = TRUE,
  use.constraint.fn = FALSE)

Arguments

raster.layer
[raster::RasterLayer]

maximize
[logical(1)]

interpolate
[logical(1)] Elevation information for this type of data is just available for integer coordinates. If interpolate is TRUE the function will round the input and return the elevation for the closest integer coordinate.

use.constraint.fn
[logical(1)] If TRUE the smoof function will also have a constraint function that returns FALSE for all input coordinates that are outside of the region coverd in the raster.layer. Otherwise the function will just return NA there.

Value

smoof function